requirements.txt
opencv-python
numpy
pandas
scikit-learn
odfpy
pillow
scikit-image
PyWavelets
matplotlib
seaborn
joblib